Django auth

+++ Django 13/Django_13_knowledgehub/notes/views.py
@@ -2,25 +2,27 @@ from django.http import HttpResponse, HttpRequest, HttpResponseForbidden
 from django.shortcuts import get_object_or_404, redirect, render
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth import logout
 from django.views.decorators.http import require_POST
 from .models import Note
-from .forms import NoteForm, ContactForm
+from .forms import NoteForm
 
 def home(request: HttpRequest) -> HttpResponse:
     return render(request, "notes/home.html")
 
-
+@login_required
 def notes_list(request: HttpRequest) -> HttpResponse:
-    notes = Note.objects.select_related("author", "category")
+    notes = Note.objects.filter(author=request.user).order_by("-created_at")
     return render(request, "notes/notes_list.html",
                   {"page_title": "Notes List", "notes": notes})
 
-
+@login_required
 def note_detail(request: HttpRequest, note_id: int) -> HttpResponse:
-    note = get_object_or_404(
-        Note.objects.select_related("author", "category").prefetch_related("tags"), pk=note_id)
+    note = get_object_or_404(Note, pk=note_id)
+    if note.author != request.user:
+        return HttpResponseForbidden("Ты типа такой хитрый что хочеш заметки других увидет? Ты хацкер?")
     return render(request, "notes/note_detail.html", {"note": note})
+
 @login_required
 def note_create(request: HttpRequest):
     if request.method == "POST":
@@ -40,6 +42,10 @@ def note_create(request: HttpRequest):
 @login_required
 def note_update(request: HttpRequest, note_id: int) -> HttpResponse:
     note = get_object_or_404(Note, pk=note_id)
+
+    if note.author != request.user:
+        return HttpResponseForbidden("Ты типа такой хитрый что хочеш заметки других изменить? Ты хаткер?")
+
     if request.method == "POST":
         form = NoteForm(request.POST, instance=note)
         if form.is_valid():
@@ -52,6 +58,11 @@ def note_update(request: HttpRequest, note_id: int) -> HttpResponse:
 @login_required
 def note_delete(request: HttpRequest, note_id: int) -> HttpResponse:
     note = get_object_or_404(Note, pk=note_id)
+
+    if note.author != request.user:
+        return HttpResponseForbidden("Ты типа такой хитрый что хочеш заметки других изменить? Ты удалякер?")
+
+
     if request.method == "POST":
         note.delete()
         messages.success(request, "Note Deleted")
